refactor: utils.VertexQuery -> domain.GraphVertex
