+++ src/main/scala/domain/graph/GraphVertex.scala
@@ -1,4 +1,4 @@
-package utils.vertex
+package domain.graph
 
 import com.typesafe.config.ConfigFactory
 import domain.table.{TableList, TableName}
@@ -15,7 +15,7 @@ import domain.table.column.{
 }
 import gremlin.scala._
 
-object VertexUtility {
+case class GraphVertex(private val value: Vertex) {
 
   private val config = ConfigFactory.load()
 
@@ -28,33 +28,31 @@ object VertexUtility {
 
   /** convert to Database Table Information
     *
-    * @param vertex
-    *   [[Vertex]]
     * @return
     *   Database Table Information
     */
-  def toDdl(vertex: Vertex): TableList =
+  def toDdl: TableList =
     TableList {
       val idColumn = Map(
         ColumnName(config.getString("column_name_vertex_id")) -> ColumnType
-          .apply(vertex.id())
+          .apply(value.id())
       )
-      val propertyColumn = vertex.valueMap.map { case (key, value) =>
+      val propertyColumn = value.valueMap.map { case (key, value) =>
         ColumnName(s"$columnNamePrefixProperty$key") -> ColumnType.apply(
           value
         )
       }
       val labelColumn = Map(
         ColumnName(
-          s"$columnNamePrefixLabel${vertex.label()}"
+          s"$columnNamePrefixLabel${value.label()}"
         ) -> ColumnType.apply(true)
       )
 
       Map(tableName -> ColumnList(idColumn ++ propertyColumn ++ labelColumn))
     }
 
-  def toDml(vertex: Vertex): String = {
-    val (propertyColumnList, propertyValueList) = vertex.valueMap.unzip
+  def toDml: String = {
+    val (propertyColumnList, propertyValueList) = value.valueMap.unzip
     val valueListForSql = propertyValueList.map { value =>
       ColumnType.apply(value) match {
         case ColumnTypeBoolean   => value
@@ -66,11 +64,11 @@ object VertexUtility {
       }
     }
 
-    val labelColumn = s"$columnNamePrefixLabel${vertex.label()}"
+    val labelColumn = s"$columnNamePrefixLabel${value.label()}"
 
     s"INSERT INTO ${tableName.toSqlSentence} (${config.getString("column_name_vertex_id")}, ${propertyColumnList
         .map(columnName => s"$columnNamePrefixProperty$columnName")
-        .mkString(", ")}, $labelColumn) VALUES (${vertex
+        .mkString(", ")}, $labelColumn) VALUES (${value
         .id()}, ${valueListForSql.mkString(", ")}, true);"
   }
 }
